Add the language selector to RetroCerts pages
Move the code from the guide into a component
and add the component to the 4 retro certs pages.
